Domain Dropdown Complete
Just like the boss wanted!  A few questions remain: should domain be
editable?  Also... on Mondays that begin new payperiods, this is when we
receive our last time entries for previous payperiods (prior Sunday) but
these will never be editable or viewable in the Portal becuase as soon
as Monday starts, the portal switches over to the new payperiod.  This
okay or not?